chore: Adjust for CI
