+++ src/Uno.UI/UI/Xaml/Media/Animation/ReorderThemeTransition.cs
@@ -5,10 +5,10 @@ namespace Windows.UI.Xaml.Media.Animation;
 public partial class ReorderThemeTransition : global::Windows.UI.Xaml.Media.Animation.Transition
 {
 	[global::Uno.NotImplemented("__ANDROID__", "__IOS__", "IS_UNIT_TESTS", "__WASM__", "__SKIA__", "__NETSTD_REFERENCE__", "__MACOS__")]
-	public ReorderThemeTransition() 
+	public ReorderThemeTransition()
 	{
 		global::Windows.Foundation.Metadata.ApiInformation.TryRaiseNotImplemented(
-			"Windows.UI.Xaml.Media.Animation.ReorderThemeTransition", 
+			"Windows.UI.Xaml.Media.Animation.ReorderThemeTransition",
 			"ReorderThemeTransition.ReorderThemeTransition()",
 			LogLevel.Debug);
 	}
